+++ src/mmu.h
@@ -9,6 +9,6 @@ class MMU {
 
         void LoadROM(MMU* mmu, char *Filename);
         
-        uint16_t ReadMemory(uint16_t Address);
+        uint8_t ReadMemory(uint16_t Address);
         void SetMemory(uint16_t Address, uint8_t Value);
 };
